enriched data for articles and completed source diversity chart
