+++ client/src/helpers/Ratings.ts
@@ -21,22 +21,4 @@ export const numBiasSources = (articles: SavedArticle[], leaningOne: string, lea
 };
 
 
-export const rightBiasSources = (articles: SavedArticle[]) => {
-
-    let count: number = 0;
-
-    for (let i = 0; i < articles.length; i++) {
-        let bias = articles[i].bias;
-
-        if (bias === 'Right-Center') {
-            count++;
-        } else if (bias === "Right") {
-            count++;
-        } else {
-            continue;
-        };
-    }
-
-    return count;
-};
